move template code to its own package
